Fix a Octal bug.
For escaped strings like "x\0365x", only "\036" is the octal number, but in the original codes, it consides "\0365" (5 added) as the octal number, and invoke a NumberFormatException at 

out.write( Integer.parseInt(input.subSequence(start, end).toString(), 8) );

As we know, octal number is only 3 chars. So I add a if-section.