LANG-1093 - Add ClassUtils.getAbbreviatedName()
getAbbreviatedName() allows to create a short version of a class name.
It accepts a parameter which gives the desired length, but ClassUtils will
make sure that the result keeps most of its meaning, which can result in
shorter or longer abbreviations than requested.

This pattern is applied by logging frameworks and generally accepted to produce
nice and understandable abbreviations.