LANG-877 removes unnecessary string allocation and improves hex writing.
Removes temporary allocation of char arrays and Strings from unicode escaping.
This roughly doubles the throughput of the escaping functionality mentioned in
LANG-877. As a side effect this significantly reduces garbage (for the cases
where the JVM does not allocate the char arrays / String on stack).

Note that there is a minor duplication / inconsistency which I uncovered.
CharUtils will use lowercase letters, UnicodeEscaper uppercase.