LANG-935 optimize lookup of translations by LookupTranslator
The previous implementation retrieved substrings from the input and checked if
it had an replacement for it. The problem is that this will always create
substrings (which are no longer "free" since JDK 7). This happens also for
substrings which are obviously not having a mapping.

The new implementation will no longer hash substrings, but will look for
translations that could be applied to the input.
Usually the very first character can rule out translation already, so this is
the new key for the mapping table.

This is twice as fast as the previous implementation and avoids a lot of
Substring allocation.