LANG-1248 FastDatePrinter Memory allocation regression
When the code was migrated from StringBuffer to Appendable in LANG-1152.
We've lost the ability to modify the buffer (setCharAt) 
The new implementation of appendFullDigits allocate a temporary char
array to work around that limitation.
This is a major source of memory allocation which is not present in
version 3.4.
Test case : 
Single fastdateformat instance
100000 format of long 
pattern : yyyy-MM-dd'T'HH:mm:ss.SSSZ
lang 3.4 : ~30 MB allocated
lang 3.5-snapshot : ~40 MB allocated
-> ~9.3 MB are coming from appendFullDigits called from
PaddedNumberField

This commit add a fast path for 1 to 4 digits which avoid the memory
allocation from the temporary work array.