Apply JavaDoc checkstyle only to public methods
The defacto style of the project requires Javadoc for public methods
only, but the checkstyle Javadoc check defaults to requiring them even
for private methods, generating 46 errors.

This patch sets the checkstyle Javadoc check's scope to public to clean
up the checkstyle report so it can be enabled in the CI.

If we wish to reset the check to a laxer scope, the aforementioned
errors should be fixed first.