Upgrade maven-checkstyle-plugin to 2.17
This patch upgrades maven-checkstyle-plugin to the latest available
version, 2.17.

This is done in order to consume a fix for checkstyle wrongfully
reporting an error if the @return javadoc tag was used in an
annotation type, as it is in Guarded (line 36).

Note that checkstyle has removed the RedundantThrows check (see
discussion at https://github.com/checkstyle/checkstyle/issues/473),
so it was removed from the project's checkstyle.xml configuration.