Add checkstyle to Travis CI (closes #254)
Currently, checkstyle is only run as part of the reporting phase, and
it's up to the developer to check the report manually.

This patch adds the checkstyle configuration to the build plugins so
it can be used to check the code (as opposed to just generate a
report of the failures) and adds it to Travis CI's configuration so
every new patch will be automatically checked against it.