Remove inequality check from shuffle tests
The ArrayUtils.shuffle methods could conceivably shuffle an array in
a way that leaves the shuffled array equal to the original array, and
thus asserting that the shuffled array differs from the original
array in the test only passes statistically.

The chance of this happening increases as an inverse of the number of
distinct values in the array. E.g., in the edge case of an array
where all the elements are equal to each other, shuffling the array
has a 100% chance of producing an array that's equal to the original
array. A less extreme case is the case of
ArrayUtilsTest#testShuffleBoolean that shuffles an array that contains
ten elements, but only two distinct values (true and false) has been
seen to fail in Travis CI [1].

This patch removes this problematic assertion and leaves only the
tests on the content of the array.

[1] https://travis-ci.org/apache/commons-lang/jobs/346806985