[LANG-1776] Use GitHub URL in `project.scm.url` (#1405)
* [LANG-1776] Use GitHub URL in `project.scm.url`

This PR updates the `project.scm.url` field in the Maven POM to point to the GitHub repository URL (`https://github.com/apache/commons-lang`).

While this field is not used directly by Maven during builds, it is increasingly leveraged by tools like *Dependabot* to:

* Associate the Maven artifact with its GitHub repository
* Automatically include release notes, changelogs, and commit history in pull requests for new versions

By using the GitHub URL, we improve the quality and completeness of metadata shown in automated upgrade PRs and make it easier for users to trace changes between releases.

* fix: Add `ciManagement` to point at GitHub

This adds a `ciManagement` property that points to the GitHub Actions URL. This is a valid alternative to `project.scm.url` to allow Dependabot and other tooling to associate an artifact with the GitHub repo.