[LANG-1597][LANG-1782] Fix getMatchingAccessibleMethod when no varargs are supplied and when the supplied varargs don't match the exact class
The current implementation discarded methods which could be potential
matches because the matching was way to strict. This change loosens the
check to allow any assignable values. The change also checks all
variable arguments instead of just the last argument.

To facilitate various number types and conversion
NumberUtils.convertIfNotNarrowing was also introduced. This allows
integers to be passed to methods which take Long... varargs