[LANG-1805] Fix inaccurate capacity calculation in join(char[])
The previous implementation of join(char[], ...) calculated the initial StringBuilder capacity based on 'array.length' instead of the actual number of elements to be joined ('endIndex - startIndex').

This caused excessive memory allocation when joining a small range of a large array (e.g., joining 5 elements from an array of 10,000).

This commit fixes the calculation to use 'noOfItems', ensuring precise memory allocation.

Signed-off-by: jher235 <tim668666@gmail.com>